task: tweaking some more
