--- internal/copy.go
@@ -1,41 +1,0 @@
-package internal
-
-import (
-	"io"
-	"os"
-	"path/filepath"
-)
-
-// copyDir copies a directory recursively from src to dst
-// src and dst must be absolute paths. This is useful to copy the
-// assets recursively.
-func copyDir(src string, dst string) error {
-	return filepath.Walk(src, func(path string, info os.FileInfo, err error) error {
-		if err != nil {
-			return err
-		}
-		dstPath := filepath.Join(dst, path[len(src):])
-		if info.IsDir() {
-			return os.MkdirAll(dstPath, info.Mode())
-		}
-
-		srcFile, err := os.Open(path)
-		if err != nil {
-			return err
-		}
-		defer srcFile.Close()
-
-		dstFile, err := os.Create(dstPath)
-		if err != nil {
-			return err
-		}
-		defer dstFile.Close()
-
-		_, err = io.Copy(dstFile, srcFile)
-		if err != nil {
-			return err
-		}
-
-		return os.Chmod(dstPath, info.Mode())
-	})
-}
